requirements.txt
Flask
python-dotenv
langchain-openai
pydantic